This commit automatically drops distributions from Chapel codes on the XMT.
This commit automatically drops distributions from Chapel codes on the XMT.
If the parser determines that a user is trying to compile some code which
contains 'dmapped ...,' we automatically ignore the '...' s.t. the user
didn't include it. The XMT is seen as a single-locale machine; the
distribution is moot.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17582 3a8e244f-b0f2-452b-bcba-4c88e055c3ca