Added the stm equivalent of chpl_stm_fork_fast. The transactional version is similar in that it does not spawn a separate task on the remote locale to execute the forked function.
This is required in RA2 kernel, for the same reasons listed in the 
HPCC report for the RA kernel. In the RA2 kernel, we have an
on-construct within the atomic block and we use the local block to
assert that we are not doing any communication within the on construct.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17623 3a8e244f-b0f2-452b-bcba-4c88e055c3ca