Added a simple contention management alogrithm based on random backoff. This algorithm executes a do-nothing loop for a random number of iterations. We use the number of retries to fix an upper limit on the number of iterations.
I am seeing real benefits even with such a simple policy on 
a workstation, esp. when we have high contention, compared to blindly
retrying the transaction when it aborts. 

It should be interesting to see how CHPL_TASK_SLEEP works out for
this purpose. Also, this algorithm does not differentiate between
local and remote transactions. Ideally, we want a policy that gives
higher priority to remote transactions. But that would require a
conflict resolution scheme which we don't support today.  



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17630 3a8e244f-b0f2-452b-bcba-4c88e055c3ca