Added a simple sed script that replaces certain STM macros with a set of dummy macros that basically perform the intended operation directly or using other Chapel runtime macros.  For example, the CHPL_STM_LOAD_REF(tx, dst, src...) macro is replaced by CHPL_DUMMY_STM_LOAD_REF and is implemented as:   dst = *(src)
The primary motivation here is to determine how the compiler generated STM
code compares against the bare-minium STM instrumentation that is required. 

Note: The script is by no means smart (or intended so). It is basically tailored
for the RA and RA2 kernel and replaces one or more occurrences of the STM_LOAD_REF 
incorrectly in the generated code of these kernels. So a bit of hand-modifcation
is still required. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17637 3a8e244f-b0f2-452b-bcba-4c88e055c3ca