Fix commit made in r17327.
Commit #17327 was supposed to allow type arguments to be passed to
functions.  I must have missed this in the original commit (or maybe
things have changed?), but the code was actually passing a temp
variable of that type to the extern function (added during the
normalize pass in insert_call_temps()).

This commit fixes this by replacing that temp variable with a SymExpr
that points to the TypeSymbol.

Added a few more cases for the sizeof() test now that config types are
supported.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17650 3a8e244f-b0f2-452b-bcba-4c88e055c3ca