Added an optimization that replaces the calls to CHPL_ASSIGN_SVEC with element-wise assignment.  There is a limit on the size of homogenous tuples considered for this optimization.  This is a peep-hole optimization that happens during codegen.
Background:

CHPL_ASSIGN_SVEC (introduced in r16448) is used for assignment of
homogenous tuples which are implemented as static C arrays.  Because
CHPL_ASSIGN_SVEC calls memcpy(), we observed a huge performance hit on
the arrayPerformance test after r16448.  Some of this was regained
after improvements to scalar replacement (r17289), but we miss a lot
of opportunities for scalar replacement due to such tuples being
within other structures (i.e., records/classes).  In these cases we
don't replace the tuples (though we could if we had better analysis).

New flags:

      --[no-]tuple-copy-opt              Enable [disable] tuple (memcpy)
                                         optimization
      --tuple-copy-limit <limit>         Limit on the size of tuples
                                         considered for optimization

The default limit is actually set to the scalar replacement limit.
Neither of these is well tuned.

Test changes:

Changed compiler help output.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17768 3a8e244f-b0f2-452b-bcba-4c88e055c3ca