When a open with write is requested it can cause problems when used from several locales in parallel, as they would create the file again and again.
To solve that the file must be opened readwrite, but then a prior open
with write would be needed when creating a new file. 

To avoid the need to do a prior write, I have changed the fileaccess
mode internally, so once created it is reopened automatically in
readwrite mode, so it can be used from several locales at once.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/pario@17774 3a8e244f-b0f2-452b-bcba-4c88e055c3ca