Fixed the use of the timing variables.
FYI: This version is about 3 orders of magnitude slower than the
original C version.  The binary splitting phase (bs) is the most time
consuming portion in both versions.  I believe the Chapel version is
suffering from our somewhat inefficient implementation of 1-D arrays
used to implement the stacks in this algorithm.  Every access to the
stack requires accessing the descriptor and computing offsets and
finally calling the _ARRAY_GET macro.  These accesses are passed by
reference (inout) to the GMP routines and thus cannot be cached into a
temp variable in the Chapel program.  I tried using the "get ref"
primitive to capture the reference, but it wasn't quite able to do the
trick (or I wasn't using it correctly).  Regardless, this isn't
something we'd want the user to have to do.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17780 3a8e244f-b0f2-452b-bcba-4c88e055c3ca