Pushing runtime generated files down one directory based on comm. substrate
In order to support multiple Chapel runtimes to coexist
simultaneously, this change pushes the communication substrate into
the path used for generated files for the runtime (.o's, Makefile
dependences, .a's).  This path will show up in the gen/ subdirectories
under runtime and in the lib/ directory under $CHPL_HOME.  The net
effect is that I should be able to build a copy of the runtime for
each of CHPL_COMM_SUBSTRATE mpi and CHPL_COMM_SUBSTRATE portals (say)
and switch between them simply by changing my CHPL_COMM_SUBSTRATE
environment variable without recompiling the runtime.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17799 3a8e244f-b0f2-452b-bcba-4c88e055c3ca