Improving the logic of the default launcher for XT/XE platforms
This change improves the logic for choosing a default launcher for
XT and XE platforms by queueing off of (a) the target platform and
(b) the presence/absence of qsub and/or aprun.  

In our previous releases, we hard-coded the launcher to be aprun
when using the portals conduit (a roundabout way for detecting
that we were on XT/XE) and then hard-coded it to be pbs-aprun
within the Chapel module built for XTs.  This caused problems when
someone wanted to use the mpi conduit and/or used a system that
didn't support qsub and/or aprun.  Thus, this change makes logic
related to selecting aprun and/or qsub based on targeting the XT
or XE rather than the substrate, and chooses the default based on
the following logic:

If aprun not in the path: launcher = none
If aprun in the path
  If qsub in the path:    launcher = pbs-aprun
  else                    launcher = aprun

This should permit Ryan to set one less environment variable in the
module and make the code more portable to (or rather, less broken on)
systems that have different launching mechanisms than ours.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17801 3a8e244f-b0f2-452b-bcba-4c88e055c3ca