Implementation of parallel iterators for associative arrays and domains.
This check-in enables:

- parallel iteration over associative arrays and domains
- parallel whole array assignment of associative arrays
- parallel reductions over associative arrays

Associative domains are still not thread-safe and thus whole-domain
assignment is still serialized.

This is a simple parallelization that hands out equal-ish sized chunks
of the hash table to each task, without regard for the number of
indices within each chunk.  This is fast and cheap, but it requires
that the follower domains match the leader's domain precisely.  A
quick initial check is performed to see if the domain's are the same
and if not, a O(tableSize) check is performed to see if the domains
match.

I made an attempt at parallelizing by the existing indices for better
load balancing (and zippering with different domains).  I was able to
get a little speed up by caching locations in the table, but the
additional work was not worth it given the high overheads of our
current implementation of 1D arrays and iteration over them.

Added the beginnings of a associative domain robustness suite:

      basic: promotion, reduce, whole-array/domain assignment, zipper, etc.
      stress: create lots of domains and arrays
      performance: track parallel performance of domain and array iteration

Other test changes:

- Updated test output for tests that were using whole-array assignment

- Changed whole array assignment to domain iteration for operations
  over arrays with different domains

- associative/waynew/indef8-fail2.chpl: Split this into 3 tests.

    indef8a.chpl:
    Assignment between arrays of the same shape allowed (as per r9208)

    indef8-fail1.chpl:
    Serial index out of bounds error with explicit indexing into the array

    indef8-fail2.chpl:
    Parallel zippered iteration with different domains

- studies/madness/aniruddha/madchap/recordBug.chpl: Serialized a
  reduction due to arrays of different domains being copied




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17803 3a8e244f-b0f2-452b-bcba-4c88e055c3ca