This commit changes localSlice calls to dsiLocalSlice.
This commit changes localSlice calls to dsiLocalSlice. The user would still
call the code via localSlice, but this would in turn call the correct
dsiLocalSlice distribution code. This change also starts work on creating
the default arithmetic array implementation for dsiLocalSlice. It first
confirms the user is asking to slice locally. Currently, it then just halts
as 'unimplemented code,' however the bones are in place: create a new
DefaultArthmeticDom using the stride of the passed in ranges, sets the new
domain's ranges value, and passes control to dsiSlice. What remains to be
done is unification of a return value: wrapped in a record or not. The
working code has been tested using GASNet over multiple locales to confirm
the right error message comes out. The code is:
var A:[1..10] real = [i in 1..10] i; writeln(A.localSlice(1..10 by 2));
on Locales(1) do { writeln(A.localSlice(1..10 by 2)); } Right now, the
first writeln doesn't work given the unimplemented status of the code, but
if you comment that out, the second writeln does the correct thing.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17807 3a8e244f-b0f2-452b-bcba-4c88e055c3ca