Added additional specialization to the serial iterator for performance reasons.
This change specializes the code for non-stridable arrays and gets
performance for 1-d non-stridable arrays (arrayPerformance.chpl) back
to before r17221.

BACKGROUND:

In r17221, David made the iterator work for arrays over domains of
negative stride.  Unfortunately, this seemingly incuous change
disables the "single loop iterator" optimization.  This was seen in
the arrayPerformance benchmarks.  It's possible that we could further
specialize if we can determine the sign of the stride at compile time
and/or have the user assert the stride to help with the optimization.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17816 3a8e244f-b0f2-452b-bcba-4c88e055c3ca