Added support for localSlicing for DefaultArithmetic arrays
This change builds on the background work Lee did and adds
support for localSlicing to DefaultArithmetic arrays.  It
also adds some test to be sure that it works OK.

When Lee was doing his initial change, part of our surprise was to
find that dsiSlice returns a layout/dist array class while
dsiLocalSlice returns an array value class (the _array record).  This
seemed asymmetric and worth fixing, but it turned out to be for a
reason: because slicing returns an alias to the original array, in the
localSlice case, the array you want to establish an alias for (and
therefore bump the reference count of) is the appropriate local one
within the local array class, not the global one.  For example, in a
Block distribution, when a localslice is taken, we don't alias the
whole global array that's spread across multiple nodes, we simply
alias one of the local arrays within one of the locale's descriptors.
Since ChapelArray.chpl has no idea what representation is used within
a distribution's classes, it has no real choice but to have the
distribution (or for symmetry, the layout) create the array class and
do the reference counting/aliasing itself.

Net result: the dsiLocalClass should return the array wrapper record.

The challenge for DefaultArithmetic arrays is that they don't store
their data using a Chapel array, but rather using a ddata object which
doesn't support creating aliases.  So I implemented a special
localSlice() function in ChapelArray.chpl to grab the local slice and
do the reference counting there.  This felt cleaner than putting code
related to wrapper records and reference counting into the
DefaultArithmetic class (and in fact seemed like the only way to do it
without changing the interface given that you can't walk from a class
back up to its wrapper record at runtime).

I have not yet done the acid test which is to see whether this permits
fft to work when compiled for the XMT.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17819 3a8e244f-b0f2-452b-bcba-4c88e055c3ca