Guarded calls to halt("...can't be zippered with...") with a sync variable.
Motivation: in general, when multiple threads invoke halt() simultaneously
when running with CHPL_COMM=gasnet, it is often the case that the program
does not terminate properly.

That's because halt() calls chpl_error(), which eventually calls
chpl_comm_exit_any_dirty(). The latter runs pthread_join() (but only
if the caller of halt() is on Locale #0) and gasnet_exit() (all locales).
So if halt() is called from Locale #0 and also from another thread,
the other thread may execute gasnet_exit() prior to the first thread
executing pthread_join(), so pthread_join() fails.

This scenario has not been highlighted by out tests (except for one
which is labeled as a future) until now that we have parallel sparse
iterators. These iterators can do that when zippering different domains,
in which case each follower task invokes halt("...can't be zippered with...").

So to prevent this scenario from happening in conjunction with these
iterators, we guard their calls to halt with a sync variable.
The first iterator to invoke halt() locks the variable (and never releases it)
so the others are stopped waiting to acquire the lock (and never call halt()).

-----------
See also: the thread ''SF.net SVN: chapel:[17018] trunk/test''
on chapel_dev@cray.com mailing list
(majordomo/lists/chapel_dev-2010/18083.html et al.).

-----------
For the record, here is what we also tried as a more general solution.

We tried to add '_haltLock = 1' *inside* halt()
(prior to calling chpl_error), just like we have it *prior* to halt()
in LayoutCSR.chpl and DefaultSparse.chpl in this check-in.

Here are the issues we encountered (and the solutions due to Brad):

- Since halt() is in ChapelIO.chpl, it was natural to define the _haltLock
variable in ChapelIO.chpl as well. However, that variable's definition
was not processed by the compiler by the time it was processing some
of calls to halt(), so that resulted in a compilation error.

- We moved the variable definition to ChapelBase.chpl. The above compilation
error went away. However, this time we were unable to compile chpl_rt_utils.c
during 'make runtime'. This is because passes/codegen.cpp does not generate
C++ declarations for Chapel variables in ChapelBase.chpl upon --runtime.

- Finally we left _haltLock in ChapelIO.chpl, but modified ChapelStandard.chpl
so that ChapelIO.chpl came right after ChapelBase.chpl. All compilations
proceeded successfully.

- However, during the standard pre-checkin testing on the chapXX farm,
the following tests failed:

 test/arrays/deitz/part7/test_descriptor_frees.chpl
 test/memory/figueroa/LeakedMemoryArrayOfClasses.chpl
 test/memory/figueroa/LeakedMemory2.chpl

because memoryUsed() reported the difference of 144, 16, and 16, resp.
with out change.

Because of the latter, I tabled this change for now and went
for a more specific solution, which does not leak memory.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17821 3a8e244f-b0f2-452b-bcba-4c88e055c3ca