Fixing localslice and adding dsiIndexLocale to DefaultArithmetic
In yesterday's change to localslice, I made a mistake and used a slice
of the array's domain to create a new domain rather than creating a
DefaultArithmeticDom which is what I thought I was doing.  This caused
problems when using true distributions (like Block) on multiple locales
(which I did not test) because the slicing domain created was not local
and therefore violated the local block that the localslice may have been
part of (which is why FFT failed).  This change fixes that issue.

It also adds a dsiIndexLocale method to DefaultArithmetic which I believe
should permit FFT to now compile all the way through on XMTs (I was able
to get to the C compile step on a linux workstation by setting CHPL_TARGET_PLATFORM
to xmt).

Added a test to lock in the dsiIndexLocale behavior and also to test a few
out of bounds cases for local slicing.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17831 3a8e244f-b0f2-452b-bcba-4c88e055c3ca