Changed the meaning and computation of the build version
Traditionally, we've used a 3-number version number where the third
number -- the build version -- was an indication of the number of SVN
commits that had occurred since the last release.  The implementation
of this required us to update the 'updateBuildVersion' script on every
release to reset the value to subtract from the SVN revision number to
make the release zero.

Recently, we switched to a 4-number version number where the fourth
number became the build version.  As part of that change, I wanted to
simplify how the build version is computed and printed.  The new
scheme is as follows:

1) When we build a release, we drop a BUILD_VERSION file into it that
   contains the value 0.

2) When a developer builds off of the SVN tree, the Makefiles update
   the BUILD_VERSION file to reflect the current SVN revision number.

3) When the compiler is executed with --version, it only prints out
   the build version if it is non-zero or being run as a developer
   (via --devel/CHPL_DEVELOPER).
   
The net effect of this should be that anyone working off of a SVN tree
should get a 4-digit version number while anyone working from the
release should get the 3-digit number (unless they are using the
release in developer mode in which case they'll get the build version
of 0).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17832 3a8e244f-b0f2-452b-bcba-4c88e055c3ca