Capturing what I believe to be the original intent of this future
Sung pointed out that this future is passing, and reviewing the
archives, it appears it has always passed.  I believe that Steve
accidentally captured the results he was getting that day into the
.good.  Reading the log message for his commit, I've updated the
future to capture his original intent -- to generate an error message
stating that the two arrays being assigned are of different shapes.
See the notes in the .future for more details.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17842 3a8e244f-b0f2-452b-bcba-4c88e055c3ca