Implemented the lsf-gasnetrun_ibv launcher.
Basically the launcher executes this command:
  bsub ... gasnetrun_ibv ... a.out_real ...
(see launch-lsf-gasnetrun_ibv.c for details).
It works on the cx1 and cx1000 machines I have available.

Open issues:

1) 'bsub' generates this additional output:
     Job <705> is submitted to default queue <normal>.
     <<Waiting for dispatch ...>>
     <<Starting on c10>>
which I don't know how to suppress. This confuses tools
that compare program output against the expected.

It may be possible to change that by using a non-interactive mode
(e.g. -K instead of -I; in conjunction with -i/-o/-e)
but that will hide from the user any indication of progress.

2) By virtue of the bsub option I am using to request
a single process per node (i.e. -R "span[ptile=1]")
if there are fewer nodes than the number of locales that
the user requests with the -nl arg, bsub hangs in "waiting for dispatch".


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17864 3a8e244f-b0f2-452b-bcba-4c88e055c3ca