Various typo fixes and rephrasings.
Replaced "elide" with "omit" (by a user's request (George Sigut),
to simplify English usage).

The spec claimed that the combination of int(64) and uint(64) was illegal
as arguments to bitwise and, or, xor. The actual implementation
was changed allow these combinations (see ChapelBase.chpl).
I updated the spec to correspond to the implementation
and added rationale for why we allow mixing an int and a uint.

Spread operator names in the 'operator-name' syntax list, using $ $ $ $.
(A prettier solution would be great.  \mbox did not seem to work
in the syntax environment.)

Wrote definitions of operators on tuples in English. The previous
definitions (by Steve) were given via advanced Chapel code. It is
unclear why Steve chose to do it this way - his commit message
merely nods to a spec review (see r16713). But to me (and Brad)
this seemed like a too-heavyweight approach to conveying
a couple of simple points (equal tuple sizes, result types
based on component-wise application of the operator).

One notable exception to the above is that code-based definitions for
>, >=, <, <= on tuples are more precise than English (especially in
terms of exactly which comparisons on arguments' components they
execute) and are fairly easy to understand. However, Brad and I agreed
that this exposes too much implementation detail in the spec. So
the only implementation detail I exposed is that these operators
can be implemented as short-circuiting.

In the grammar (and accompanying text) renamed 'formal-tag' to 'formal-intent'
because 'param' and 'type' are called "intents" in the Generics section.
Also, marked 'formal-intent' as optional in a 'formal'.

Allowed 'empty-statement' as a 'record-statement' or a 'union-statement',
in accordance with the implementation and to match 'class-statement'.

range.order -> range.indexOrder

Clarified the default constructor story based on what we have at the moment.

Moved the requirements on constructors for generic classes into the generics
chapter and added examples.

Removed the test_ prefix from auto-generated test filenames in Classes.tex.

Much of the above includes contributions from Brad.
Some of the changes I am checking in are work in progress.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17875 3a8e244f-b0f2-452b-bcba-4c88e055c3ca