Added a new environment variable (CHPL_ASYNCPE_USE_DYNAMIC) to enable/disable dynamic linking for Cray PEs (xt-asyncpe module).  This was a last minute addition due to issues run into by Srini on jaguar where dynamic linking is not supported.
If true, pass the -dynamic flag.  If false, pass the -static flag.
Otherwise, use the default (currently static linking).

The default is to *not* set the variable.  We will override this in
the XE module.

Modified printchplenv to print out the variable's value if it's
relevant (i.e., if we are using a Cray PE and it's set to true or
false).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17906 3a8e244f-b0f2-452b-bcba-4c88e055c3ca