Small updates to environment section of man page
Fixed a reference to printchplenv.pl to reflect the lack of script
extensions now.

Was looking to see whether or not to add a CHPL_TIMERS entry to
the man page and decided not to given the small number of choices
and fact that the default is well-specified.

Similarly, pulled the description of CHPL_THREADS since most users
can think in terms of CHPL_TASKS (and it will default intelligently).

Updated the wording to clarify that this isn't a complete set of
environment variables, but simply some of the more important ones.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17920 3a8e244f-b0f2-452b-bcba-4c88e055c3ca