This change adds the XMT pragma noalias to follower iterators.
This change adds the XMT pragma noalias to follower iterators. To use it,
issue __primitive("noalias pragma") inside the loop in question. expr.cpp
will place the line _Pragma("mta noalias <var>") for each of the variables
inside the block the primitive resides within. Note that every variable is
picked; in the future we will only generate this for invariant pointers
(and we will also likely attach it automatically to the leader iterator).
A Makefile change was required to suppress the warnings for every variable
that generating a noalias for is moot or plain dumb. Note that
repositionDefExpressions.cpp changes, too. At the end of this pass, we
eliminate all the multiple instances of noalias pragma from a given block,
and we move the code to just before the for loop.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17935 3a8e244f-b0f2-452b-bcba-4c88e055c3ca