Renamed util/devel/README.devel to README, to match our current practices. In more detail (from Brad):
Typically we use the name README.devel in cases where there's a README 
that we want to include in the release as well as other notes that we only 
want to be for developers.  The util/devel directory is not intended for 
release in general, so this README could simply be called 'README'.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17944 3a8e244f-b0f2-452b-bcba-4c88e055c3ca