Language spec: changed the typesetting of 'nil' in subsection headings to be similar to how we typeset 'this' and 'these' in nearby subsections.
Motivation: Michael Ferguson reported seeing the following fragment
verbatim in his evince viewer (in the table of contents):

  The [language=chapel,basicstyle=,keywordstyle=]!nil! Value

while we can see this entry as intended (i.e. 'The nil value', with
nil typeset as a Chapel keyword) in our evince and pdf browsers.

My guess is that Michael's browser is old whereas the above snippet
exercises a feature of a newer pdf format version.

The change being committed should make the resulting PDF viewable
properly in browsers like Michael's.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17954 3a8e244f-b0f2-452b-bcba-4c88e055c3ca