Added ability for launchers to print out launcher-specific help
As a background activity today, implemented a new hook in the launcher
interface that permits launchers to print out their own help
information prior to the generic help message for a Chapel executable.
For most launchers, I currently stubbed this out to a no-op, but I
filled in help information for the mtarun and marenostrum launchers
(the two that currently take command-line arguments).  Next time I
work on an XT/XE, I'll probably bring those up-to-snuff if somebody
doesn't beat me to it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17964 3a8e244f-b0f2-452b-bcba-4c88e055c3ca