Small mods to how SVN part of version number is calculated
Just before the last release, I updated the Chapel version number to a
4-part version number where the 4th part represents the SVN revision
number.  When I was describing the logic at a meeting, one of the
things that we realized was that I still had a "if CHPL_DEVELOPER is
set" aspect to it which didn't seem relevant -- anyone who is building
a Chapel compiler from the .svn sources should get the build part of
the version number.  This small change modifies the Makefile and the
script used to update the build number so that it will update the
BUILD_VERSION file (which represents that fourth part of the version
number) if (1) $CHPL_HOME/.svn exists, and (2)
$CHPL_HOME/util/devel/updateBuildVersion exists.  It no longer matters
whether or not CHPL_DEVELOPER is set.

In a release, neither of these conditions apply, so the release
tarball should never try to update its BUILD_VERSION file which is set
to 0 in the gen_release script.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17968 3a8e244f-b0f2-452b-bcba-4c88e055c3ca