Moving location of compVersion file to a slightly better location
In my last commit I moved updateBuildVersion from buildRelease/
to devel/ because it's actually called in every build of the
compiler (from an SVN repository), not just when building the
release.

Similarly, the compVersion script is used in every build of the man
page, not just the release, so I moved it for similar reasons.  At
this point I think util/buildRelease only contains scripts that are
used to build the release.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17969 3a8e244f-b0f2-452b-bcba-4c88e055c3ca