Add an interface for non-blocking get operations for the Chapel runtime libraries.
This commit adds an interface for non-blocking gets that may be
utilized if the communication layer and the tasking layer have the
necessary support.  A communication layer (e.g., GASNet) would support
a non-blocking get operation as well as a way to test or wait for the
completion of the operation.  A tasking layer (e.g., Nanos or
Qthreads) would implement a get operation based on the non-blocking
support exposed by the communication layer.

Specifically, the communication layer would #define:

   CHPL_COMM_HAS_NB        defined if non-blocking comm supported for runtime
   CHPL_COMM_GET_NB()      *** required ***
   CHPL_COMM_TEST_GET_NB() *** required ***
   CHPL_COMM_WAIT_GET_NB() *** optional ***

The tasking layer would #define:

   CHPL_TASK_COMM_GET()    uses the comm layer routines to implement
                           non-blocking gets

In chplcomm.h, if CHPL_TASK_COMM_GET() is defined, it is used to
implement CHPL_COMM_GET(), otherwise the normal blocking get is used:

          #ifdef CHPL_TASK_COMM_GET
          #define CHPL_COMM_GET(local, wide, type, ln, fn) \
            CHPL_TASK_COMM_GET(local, wide, type, ln, fn)
          #else
          #define CHPL_COMM_GET(local, wide, type, ln, fn) \
            CHPL_COMM_BLOCKING_GET(local, wide, type, ln, fn)
          #endif

In a tasking layer specific header file, if CHPL_COMM_HAS_NB is
defined, define CHPL_TASK_COMM_GET() and implement it using the
non-blocking communication macros.

The non-blocking interface for the GASNET comm layer has been
implemented in this commit.

Also, added a function CHPL_PER_PTHREAD_TASKING_INIT() to the tasking
layer that should be implemented by the tasking layer to do any thread
specific initialization for threads that are not started by the
tasking layer itself.  For example, it is used in the gasnet comm
layer when the polling thread is started.

In the future, the non-blocking communication interface could also be
used directly by the compiler to overlap communication and
computation.  Care would need to be taken to insure deadlock-free
operation for comm layers that do not support non-blocking
communication.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17970 3a8e244f-b0f2-452b-bcba-4c88e055c3ca