Small fix to yesterday's version number change
One of the things yesterday's change did was to *not* create a
BUILD_VERSION file if the user was not working from an svn context.
This causes problems when you 'make' from an svn export'ed copy of the
tree, which our opts testing does (which is why it showed up empty
last night) because our compiler needs a BUILD_VERSION file to exist
when we build it.

The fix is easy, though and cleans up a piece of gen_release as well.
In particular, if the Makefile isn't going to query SVN to determine
BUILD_VERSION it simply creates a BUILD_VERSION file with "0" in it
(which is the cue to the compiler code that it shouldn't print out the
fourth part of the version number).  Previously we had been having the
gen_release script create this file which spread this issue out over
too wide a code base.  The result is much more elegant on both sides
and should re-enable the opts testing tonight.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17971 3a8e244f-b0f2-452b-bcba-4c88e055c3ca