Adding a new developer flag --break-on-id to quickly locate AST creation points
This change (co-coded with Lee and Vass) adds a new developer flag
named --break-on-id that takes an integer argument <id> and calls
gdbShouldBreakHere() when the AST with that id is created.  When
running in --gdb this gives you a quick way to find when a certain
AST node was created.  While this can be done with a conditional
breakpoint in gdb, that tends to be much slower and more painful
(in my experience).  Using this flag outside of the --gdb flag does
nothing since gdbShouldBreakHere() is a no-op.

A nice future change might be to have this flag take a list of IDs,
but that would require a string argument and parsing it which we
didn't undertake here.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17972 3a8e244f-b0f2-452b-bcba-4c88e055c3ca