Add support for thread and task layer specific module code.
This commit enables the threading and tasking layers (defined by
CHPL_THREADS and CHPL_TASKS) to define layer specific module code.

There are 2 empty stub files in $CHPL_HOME/modules/internal for this
purpose: ChapelThreadsInternal.chpl and ChapelTasksInternal.chpl.  If
a threading or tasking layer needs layer specific module code, it
should be placed in a file by the same name but in a subdirectory of
$CHPL_HOME/modules/internal named by $CHPL_THREADS or $CHPL_TASKS.
Other layer specific files can also be placed in those directories
(e.g., to be included by the 'use' statement).

The internal module search path is now prepended with
$CHPL_HOME/modules/internal/$CHPL_THREADS and
$CHPL_HOME/modules/internal/$CHPL_TASKS.  To avoid unnecessary
warnings, the warning message that states that the module name was
found in multiple directories is disabled for internal module search
only.

Test changes:

test/functions/diten/compilerWarningDepth.good:
- Update new line numbers in the ChapelBase module



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17973 3a8e244f-b0f2-452b-bcba-4c88e055c3ca