// two independent changes, combined to reduce testing
David Iten's utility to print the number of each type of primitive in the AST.
  include/primitive.h
  main/runpasses.cpp
  AST/primitive.cpp

David writes: "If you uncomment the commented out line in runpasses.cpp
[that calls printPrimitiveCounts() -ed.] it will walk the AST between
passes and print the number of each type of primitive encountered.
This could be useful as a developer compiler option."

---

Fixed the issue reported by Christopher Swenson on 11/03/2010.
  passes/scopeResolve.cpp

The issue is that the compiler failed to report an error on the following code:
  class A { var b; } var x = new A();

(Since 'b' is a generic field, its initializing value must be passed
to each constructor of A. The checking on param and type fields was
already done properly.)

Turns out that the default constructor was getting the default expression
for 'b', which made A() (no actuals) legitimate. Fixed in build_constructor()
by strengthening the check for when to create a default expression ('init'
in the code). (Also did minor code refactoring and variable renaming there.)

Changed/added tests for this fix:
  test/classes/vass/generic-var-const-field-error.future
  test/classes/constructors/*


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17979 3a8e244f-b0f2-452b-bcba-4c88e055c3ca