Minor changes to task parallel codes
On the way in today I realized that after patting myself on the back
for getting these working in a way that generated the same result
the other night, that I hadn't actually tested them in any interesting
configurations, such as one in which the number of iterations didn't
divide evenly by the number of tasks/locales.  So this changes those
numbers and fixes one bug that I missed.  Oops.  I believe they're
correct now.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17982 3a8e244f-b0f2-452b-bcba-4c88e055c3ca