Factored out the common code between printCallStack()-the-debugging-helper and printCallStackOnError()-user-message-printing-helper.
Moved that code into misc.cpp so it can check 'err_print' (and skip printing
the stack if it's true), which is only visible in misc.cpp.
Note that printCallStack() is not referenced in functionResolution.cpp.

The phrase ""while processing the following Chapel call chain" is needed
only in user messages, so I suppressed it for developers.

Verified that --show-calls-on-error is a user (i.e. not developer) option,
as intended.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18006 3a8e244f-b0f2-452b-bcba-4c88e055c3ca