Changing testProbSize prediffs to force recalculation of memory size
Vass and Sung recently noticed an issue in which testProbSize tests
were failing during their test runs.  It turns out that the reason was
that the computeProbSize script on which testProbSize's results are
computed is updated lazily, so if a computeProbSize.good file already
existed for a given platform, it would be used (even if it was
generated using an older version of the OS).  The nightly testing
system didn't run into this problem because it starts from a clean
version of the source tree each night.

I fixed the issue by simply forcing the recomputation of
computeProbSize.good when these tests are run.  This increases the
possibility of the race between these three test directories mentioned
in previous commits (where multiple tests on a single platform may try
to regenerate computeProbSize.good simultaneously), but the chances
are slim enough that I didn't bother undertaking a fix for it (like
having the testProbSize tests generate the computeProbSize.good output
file to their own local directories) tonight.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18009 3a8e244f-b0f2-452b-bcba-4c88e055c3ca