Filing a bug to capture the question about support for param complex values
This is a capture of the email thread started by Paul Adamson this
week on chapel-users@lists.sourceforge.net.  I think the right thing
to do is to support param complex values, but I don't think we should
do this currently.  I've filed this as a semantic future, but given
the current specification, I think it should be added to STATUS.devel
as a bug for the next release.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18014 3a8e244f-b0f2-452b-bcba-4c88e055c3ca