start_test, sub_test: - added -sysprediff/CHPL_SYSTEM_PREDIFF   to specify a system-wide prediff script
The system-wide prediff, if specified, is invoked the same way as
other prediff scripts and intended to clean up the extraneous output
generated e.g. by the launcher on a specific platform.  Therefore it
is invoked before the directory- and test-specific prediffs and
is not invoked when the test executable is not invoked.

If we encounter challenges writing a script to filter out launcher
output but leaving the (preceeding) compiler output (when present)
in tact, we may need to elaborate on this functionality.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18026 3a8e244f-b0f2-452b-bcba-4c88e055c3ca